+++ test/main-tests.js
@@ -45,7 +45,7 @@ phantom.injectJs("./test-webpage-render.js");
 phantom.injectJs("./test-webpage-prompt.js");
 phantom.injectJs("./test-webpage-open.js");
 phantom.injectJs("./test-webpage-frames.js");
-
+phantom.injectJs("./test-webpage-callPhantom.js");
 
 var webserverTest = webServerFactory.create();
 webserverTest.listen(8083, function(request, response) {